refactor: enhance RAG file structure
