+++ rag/rag.py
@@ -1,4 +1,4 @@
-from search import Searcher
+from .search import Searcher
 
 import asyncio
 import os
@@ -41,79 +41,14 @@ async def crawl4ai_crawl_many(urls: list,crawler_config = crawler_config) -> lis
         return result   #type: ignore 
 
 
-class SlidingWindowChunking:
-    """
-    A class used to perform sliding window chunking on text data.
-    This class provides methods to split text into chunks of a specified window size
-    with a given step size. It also summarizes with an LLM if possible and filters the markdown
-    """
-
-    def __init__(self, window_size=135, step=35):
-        self.window_size = window_size
-        self.step = step
-    
-    def chunk(self, text, window_size=0,step=0) -> list[str]:
-        window_size = window_size or self.window_size
-        step = step or self.step
-
-        words = text.split(" ")
-        chunks = []
-        for i in range(0, len(words) - window_size + 1, step):
-            chosen = words[i:i + window_size]
-            if chosen:
-                chunks.append(' '.join(chosen))
-        if not chunks:
-            chunks = [text]
-        return chunks
-    
 
 
-class RAGOpenai(MasterOpenaiInterface):
-    def __init__(self, cheap_models:list[LLMModelInfo]):
-            super().__init__(cheap_models=cheap_models)
-            print(f"RAG: using {self.model}")
-            self.summarize_prompt = """You are a summarization assistant. When summarizing a text,
-              provide only a concise, clear summary without any greetings, preamble, or extra commentary. 
-              Do not include phrases like \"Sure!\" or
-              \"Here is the summary.\" Simply output the summary in a direct and succinct manner.""".replace("\n", " ")
-            self.chunker = SlidingWindowChunking(window_size=self.rate_limit//2, step=90)
 
-    async def __summarize_req(self, piece:str):
-        messages=[
-            GptMessage(role="system", 
-                content=self.summarize_prompt).model_dump(),
 
-            GptMessage(role="user",
-                content=piece).model_dump()
-        ]
-
-        [m.pop("id") for m in messages]
-        
-        completion = await self.openai.chat.completions.create( #type: ignore
-            model=self.model,
-            messages=messages #type: ignore
-        ) 
-
-        return completion.choices[0].message.content
-
-    async def summarize(self, text: str, manage_limits=False)-> str:
-        pieces : list[str] = [text]
-        n_reqs = 1
-        if manage_limits:
-            ...
-        else:
-            pieces = [text] 
-        
-        results = []
-        for piece in pieces:
-            try:
-                results.append(await self.__summarize_req(piece))
-            except Exception as e:
-                print("ERROR SUMMARIZING: " + str(e))
-        return '\n'.join(results)
-    
-
-import vdb, os
+from .rag_openai import RAGOpenai
+from .chunker import SlidingWindowChunking
+import os
+from .vdb import VecDb
 class RAG:
     def __init__(self, cheap_models: list[LLMModelInfo], brave_api_key="", TEMBO_PSQL_URL=os.environ.get('TEMBO_PSQL_URL'), top_results=4, demo_search=False):
         """This class requires a list of cheap models"""
@@ -122,7 +57,7 @@ class RAG:
         self.demo_search = demo_search
         self.chunker = SlidingWindowChunking()
         self.top_results=top_results
-        self.db = vdb.VecDb(TEMBO_PSQL_URL=TEMBO_PSQL_URL) #type: ignore
+        self.db = VecDb(TEMBO_PSQL_URL=TEMBO_PSQL_URL) #type: ignore
         self.llm = RAGOpenai(cheap_models=cheap_models)
 
     async def search_and_crawl(self, query=""):
@@ -181,13 +116,3 @@ class RAG:
 
 
     
-async def demo():
-    #rag = RAG(brave_api_key=brave_key, demo_search=True)
-    #await rag.db.retrieve_no_search("what to do in rio")
-    query = "Best Tourist attractions in Rio de Janeiro"
-    # crawled_results = asyncio.run( rag.search_and_crawl(query))
-    # segmented_results = rag.add_chunks(crawled_results)
-    # for result in segmented_results:
-    #     print("\n\n\tCHUNK BLOCK")
-    #     #[print(f"\n\t\tCHUNK: ---{chunk}") for chunk in result.chunks]
-    # return await (rag.store_search(query))
